refactor(live): one get_observations for all five venues (#288, slice 1)
Alpaca kept a parallel copy of the entire observation window pipeline. Normalising the
three naming differences, `get_observations` differed from the futures base by ONE real
line -- `limit=window_size + 50` on the fetch -- and `_default_preprocessing` by two
alpaca SDK-shape lines. Everything else was character-identical.

That copy cost three times in a row: row alignment (#395), the stale last bar (#399,
#398), and the short window (#400) each had to be pasted into both. Alpaca's half of
#397 had ZERO test coverage until #401 added a shared one -- deleting its `dropna` failed
nothing.

The split follows what the code actually needs, measured rather than assumed: of the
seven abstract hooks on the futures base, the window logic calls only three
(`_create_client`, `_validate_timeframe`, `_get_timestamp_column`). So:

- `shared/base_obs.py` -- `BaseObservationClass`: reset, __init__, get_keys,
  _default_preprocessing, _timestamps_of, _get_numpy_obs, _dummy_frame, get_features,
  get_observations. Abstract: the three above plus `_fetch_single_timeframe` and
  `_normalise_frame`.
- `shared/futures_base_obs.py` -- now 99 lines, not 361: the four kline hooks and the
  kline-based `_fetch_single_timeframe`. The four futures venues' contract is unchanged;
  they import the same name from the same module.
- `alpaca/observation.py` -- 322 -> 191 lines. It supplies a date-RANGE fetch and a
  `_normalise_frame` that undoes the SDK's (symbol, timestamp) MultiIndex.

`timeframes=` stays alpaca's public kwarg and maps to `time_frames` in `__init__`.
Renaming it would break every existing alpaca config for no behavioural gain; nothing
outside the class reads the attribute (grepped).

Two things the fold surfaced, both kept as fixes:

- Alpaca's length check only fired when BOTH arguments were lists, so
  `timeframes=[a, b], window_sizes=5` was accepted and failed later at the spec instead.
  The base checks unconditionally.
- Alpaca's `_dummy_frame` must carry `symbol`, because `_normalise_frame` drops it. That
  is the dtype-fidelity lesson from binance's fold: a dummy that is not shaped like real
  data makes `get_features` measure a frame the venue never produces.

Guards. `test_no_venue_reimplements_the_shared_base` read `vars(BaseFuturesObservation
Class)` -- one class's __dict__ -- so hoisting `get_observations` silently dropped it
from the guarded set. It walks the MRO now. And `base_exchange_tests.py`, which all five
venue test classes inherit, gains `test_this_venue_does_not_re_fork_the_shared_window_
logic`: re-forking `_default_preprocessing` onto alpaca fails it by name. A re-fork
passes every behavioural test on the day it happens, because both copies are correct at
birth -- which is exactly why it needs a structural guard.

Net -83 lines (+372/-455). Full suite: 3985 passed, 16 skipped.
